Converting Category levels models and router to typescript
